credicheck-ai/backend/requirements.txt
transformers
torch
pillow
flask
flask-cors
spacy
sentence-transformers
scikit-learn
duckduckgo-search
lime
shap
